# docs/openapi.json
{
 "components": {
  "schemas": {
...
   "AdaptiveConfig": {
    "additionalProperties": false,
    "properties": {
     "factor": {
      "default": 1.5,
      "exclusiveMinimum": 1.0,
      "maximum": 10.0,
      "title": "Factor",
      "type": "number"
     },
     "max_s": {
      "default": 86400,
      "minimum": 60.0,
      "title": "Max S",
      "type": "integer"
     },
     "min_s": {
      "default": 900,
      "minimum": 60.0,
      "title": "Min S",
      "type": "integer"
     }
    },
    "title": "AdaptiveConfig",
    "type": "object"
   },
...
   "AutowatchRequest": {
    "additionalProperties": false,
    "properties": {
     "state": {
      "enum": [
       "running",
       "paused"
      ],
      "title": "State",
      "type": "string"
     },
     "until": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Until"
     }
    },
    "required": [
     "state"
    ],
    "title": "AutowatchRequest",
    "type": "object"
   },
   "AutowatchState": {
    "additionalProperties": false,
    "properties": {
     "state": {
      "enum": [
       "running",
       "paused"
      ],
      "title": "State",
      "type": "string"
     },
     "until": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Until"
     }
    },
    "required": [
     "state"
    ],
    "title": "AutowatchState",
    "type": "object"
   },
...
   "CheckQueued": {
    "additionalProperties": false,
    "properties": {
     "queued": {
      "title": "Queued",
      "type": "integer"
     }
    },
    "required": [
     "queued"
    ],
    "title": "CheckQueued",
    "type": "object"
   },
   "CheckRequest": {
    "additionalProperties": false,
    "properties": {
     "all": {
      "default": false,
      "title": "All",
      "type": "boolean"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "force": {
      "default": false,
      "title": "Force",
      "type": "boolean"
     },
     "ids": {
      "anyOf": [
       {
        "items": {
         "type": "integer"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ids"
     }
    },
    "title": "CheckRequest",
    "type": "object"
   },
...
   "FolderIn": {
    "additionalProperties": false,
    "properties": {
     "defaults": {
      "additionalProperties": true,
      "title": "Defaults",
      "type": "object"
     },
     "is_virtual": {
      "default": false,
      "title": "Is Virtual",
      "type": "boolean"
     },
     "name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "parent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Id"
     },
     "query": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Query"
     },
     "sort_order": {
      "default": 0,
      "title": "Sort Order",
      "type": "integer"
     }
    },
    "required": [
     "name"
    ],
    "title": "FolderIn",
    "type": "object"
   },
   "FolderOut": {
    "additionalProperties": false,
    "properties": {
     "defaults": {
      "additionalProperties": true,
      "title": "Defaults",
      "type": "object"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "is_virtual": {
      "title": "Is Virtual",
      "type": "boolean"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "parent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Id"
     },
     "query": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Query"
     },
     "sort_order": {
      "title": "Sort Order",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "parent_id",
     "name",
     "sort_order",
     "is_virtual",
     "query",
     "defaults"
    ],
    "title": "FolderOut",
    "type": "object"
   },
   "FolderPatch": {
    "additionalProperties": false,
    "properties": {
     "defaults": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Defaults"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 200,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "parent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Id"
     },
     "query": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Query"
     },
     "sort_order": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sort Order"
     }
    },
    "title": "FolderPatch",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "HostOverride": {
    "additionalProperties": false,
    "properties": {
     "concurrency": {
      "anyOf": [
       {
        "maximum": 32.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Concurrency"
     },
     "min_gap_s": {
      "anyOf": [
       {
        "minimum": 0.0,
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Gap S"
     }
    },
    "title": "HostOverride",
    "type": "object"
   },
   "OnBattery": {
    "enum": [
     "normal",
     "slow",
     "pause"
    ],
    "title": "OnBattery",
    "type": "string"
   },
...
   "ScheduleConfig": {
    "additionalProperties": false,
    "properties": {
     "adaptive": {
      "$ref": "#/components/schemas/AdaptiveConfig"
     },
     "days": {
      "items": {
       "enum": [
        "mon",
        "tue",
        "wed",
        "thu",
        "fri",
        "sat",
        "sun"
       ],
       "type": "string"
      },
      "title": "Days",
      "type": "array"
     },
     "interval_s": {
      "default": 3600,
      "minimum": 60.0,
      "title": "Interval S",
      "type": "integer"
     },
     "jitter_pct": {
      "default": 10,
      "maximum": 50.0,
      "minimum": 0.0,
      "title": "Jitter Pct",
      "type": "integer"
     },
     "mode": {
      "$ref": "#/components/schemas/ScheduleMode",
      "default": "interval"
     },
     "on_battery": {
      "$ref": "#/components/schemas/OnBattery",
      "default": "normal"
     },
     "times": {
      "items": {
       "type": "string"
      },
      "title": "Times",
      "type": "array"
     },
     "window": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Window"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "title": "ScheduleConfig",
    "type": "object"
   },
   "ScheduleMode": {
    "enum": [
     "interval",
     "times",
     "adaptive",
     "manual"
    ],
    "title": "ScheduleMode",
    "type": "string"
   },
   "Settings": {
    "additionalProperties": false,
    "properties": {
     "autowatch_paused_until": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Autowatch Paused Until"
     },
     "autowatch_state": {
      "default": "running",
      "enum": [
       "running",
       "paused"
      ],
      "title": "Autowatch State",
      "type": "string"
     },
     "backlog_warning": {
      "default": 200,
      "minimum": 1.0,
      "title": "Backlog Warning",
      "type": "integer"
     },
     "browser_pool": {
      "default": 3,
      "maximum": 16.0,
      "minimum": 1.0,
      "title": "Browser Pool",
      "type": "integer"
     },
     "debug_logging": {
      "default": false,
      "title": "Debug Logging",
      "type": "boolean"
     },
     "default_schedule": {
      "$ref": "#/components/schemas/ScheduleConfig"
     },
     "default_user_agent": {
      "default": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/131.0.0.0 Safari/537.36",
      "title": "Default User Agent",
      "type": "string"
     },
     "disk_cap_gb": {
      "default": 10.0,
      "exclusiveMinimum": 0.0,
      "title": "Disk Cap Gb",
      "type": "number"
     },
     "global_proxy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Global Proxy"
     },
     "host_overrides": {
      "additionalProperties": {
       "$ref": "#/components/schemas/HostOverride"
      },
      "title": "Host Overrides",
      "type": "object"
     },
     "keep_changed_versions": {
      "default": 20,
      "minimum": 1.0,
      "title": "Keep Changed Versions",
      "type": "integer"
     },
     "notify_on_first_check": {
      "default": false,
      "title": "Notify On First Check",
      "type": "boolean"
     },
     "per_host_concurrency": {
      "default": 2,
      "maximum": 32.0,
      "minimum": 1.0,
      "title": "Per Host Concurrency",
      "type": "integer"
     },
     "per_host_min_gap_s": {
      "default": 2.0,
      "minimum": 0.0,
      "title": "Per Host Min Gap S",
      "type": "number"
     },
     "startup_delay_s": {
      "default": 30.0,
      "minimum": 0.0,
      "title": "Startup Delay S",
      "type": "number"
     },
     "static_pool": {
      "default": 32,
      "maximum": 256.0,
      "minimum": 1.0,
      "title": "Static Pool",
      "type": "integer"
     },
     "timezone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Timezone"
     },
     "toast_coalesce_s": {
      "default": 30.0,
      "minimum": 0.0,
      "title": "Toast Coalesce S",
      "type": "number"
     },
     "transient_retry_s": {
      "default": 60.0,
      "minimum": 1.0,
      "title": "Transient Retry S",
      "type": "number"
     },
     "worker_processes": {
      "default": 3,
      "maximum": 32.0,
      "minimum": 1.0,
      "title": "Worker Processes",
      "type": "integer"
     }
    },
    "title": "Settings",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
   "Window": {
    "additionalProperties": false,
    "properties": {
     "end": {
      "default": "23:59",
      "title": "End",
      "type": "string"
     },
     "start": {
      "default": "00:00",
      "title": "Start",
      "type": "string"
     }
    },
    "title": "Window",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "PageWatch engine",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/autowatch": {
   "get": {
    "operationId": "get_autowatch_autowatch_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AutowatchState"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Autowatch",
    "tags": [
     "checks"
    ]
   },
   "post": {
    "operationId": "set_autowatch_autowatch_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AutowatchRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AutowatchState"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Set Autowatch",
    "tags": [
     "checks"
    ]
   }
  },
...
  "/check": {
   "post": {
    "description": "Check a list of bookmark ids, a folder (with its subfolders), or everything.",
    "operationId": "check_check_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CheckRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CheckQueued"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Check",
    "tags": [
     "checks"
    ]
   }
  },
  "/folders": {
   "get": {
    "operationId": "list_folders_folders_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/FolderOut"
         },
         "title": "Response List Folders Folders Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Folders",
    "tags": [
     "folders"
    ]
   },
   "post": {
    "operationId": "create_folder_folders_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FolderIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FolderOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Folder",
    "tags": [
     "folders"
    ]
   }
  },
  "/folders/{folder_id}": {
   "delete": {
    "operationId": "delete_folder_folders__folder_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "folder_id",
      "required": true,
      "schema": {
       "title": "Folder Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Folder",
    "tags": [
     "folders"
    ]
   },
   "patch": {
    "operationId": "patch_folder_folders__folder_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "folder_id",
      "required": true,
      "schema": {
       "title": "Folder Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FolderPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FolderOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Folder",
    "tags": [
     "folders"
    ]
   }
  },
...
  "/settings": {
   "get": {
    "operationId": "get_settings_settings_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Settings"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Settings",
    "tags": [
     "checks"
    ]
   },
   "put": {
    "description": "Partial update: only the keys given change.",
    "operationId": "put_settings_settings_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "additionalProperties": true,
        "title": "Patch",
        "type": "object"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Settings"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Settings",
    "tags": [
     "checks"
    ]
   }
  }
 }
}